+++ Raytracer/DirectionalLight.cpp
@@ -5,7 +5,7 @@ DirectionalLight::DirectionalLight()
 	direction = direction;
 }
 
-DirectionalLight::DirectionalLight(glm::vec3 color, double intensity, glm::vec3 direction) : Light(color, intensity)
+DirectionalLight::DirectionalLight(glm::vec3 color, float intensity, glm::vec3 direction) : Light(color, intensity)
 {
 	this->direction = glm::normalize(direction);
 }
