+++ Raytracer/DirectionalLight.h
@@ -8,7 +8,7 @@ public:
 	glm::vec3 direction;
 
 	DirectionalLight();
-	DirectionalLight(glm::vec3 color, double intensity, glm::vec3 direction);
+	DirectionalLight(glm::vec3 color, float intensity, glm::vec3 direction);
 	~DirectionalLight();
 };
