+++ Raytracer/FinitePlane.h
@@ -5,12 +5,12 @@ class FinitePlane :
 {
 public:
 	glm::vec3 lengthDir;
-	double length;
+	float length;
 	glm::vec3 widthDir;
-	double width;
+	float width;
 
 	FinitePlane();
-	FinitePlane(glm::vec3 color, double albedo, glm::vec3 position, glm::vec3 normal, glm::vec3 lengthDir, double length, glm::vec3 widthDir, double width);
+	FinitePlane(glm::vec3 color, float albedo, glm::vec3 position, glm::vec3 normal, glm::vec3 lengthDir, float length, glm::vec3 widthDir, float width);
 	~FinitePlane();
 
 	Hit intersect(const Ray &ray);
